Move ResultHandler to its own file
